Move types and classes not specific to PySNMP
These classes and type definitions aren't specific for the PySNMP
back-end and are therefore moved to a new module from where they
can be shared.
